+++ app/config.py
@@ -1,0 +1,3 @@
+class Config:
+    SEND_FILE_MAX_AGE_DEFAULT = 0
+    # TEMP_DIR = 'path/to/temp/dir'  # Configuração do diretório temporário, por exemplo.
